docker-compose: add a PoC docker-compose that spins up the whole stack
This still leaves much to be desired in the details, you'll have to
have the image-buider and composer checked out in the expected locations
and you'll have to have composer certs generated in the same way as for
docker-compose of composer alone.

I also had to add a hack to the actual package.json, which I could not
figure out how to do using env vars alone.

Signed-off-by: Tom Gundersen <teg@jklm.no>